feat:新增增值服务接口定义

+++ bajiecloud-product-server/src/main/java/com/bajiezu/cloud/product/dal/entity/ValueAdded.java
@@ -45,12 +45,12 @@ public class ValueAdded extends BaseDO {
     /**
      * 续租售价
      */
-    private Integer renewalPrice;
+    private Long renewalPrice;
     
     /**
      * 划线价
      */
-    private Integer strikethroughPrice;
+    private Long strikethroughPrice;
     
     /**
      * 服务概要
